% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datSim.R
\name{datSim}
\alias{datSim}
\title{Simulated data}
\usage{
datSim(n, bp0, mu01, mu02, var01, var02, rho01, rho02, mu11, mu12, var11, var12,
  rho11, rho12)
}
\arguments{
\item{n}{sample size}

\item{bp0}{change point}

\item{mu01}{mean}

\item{mu02}{mean}

\item{var01}{variance}

\item{var02}{variance}

\item{rho01}{correltation coefficient}

\item{rho02}{correltation coefficient}

\item{mu11}{mean}

\item{mu12}{mean}

\item{var11}{variance}

\item{var12}{variance}

\item{rho11}{correltation coefficient}

\item{rho12}{correltation coefficient}
}
\value{
data.frame as (y1, y2, x)
}
\description{
Generate dara from mixture bivariate Gaussian distribution
}
\examples{
\dontrun{
#' n = 10000
bp0 = 0.8
## for x=0
rho01 <- 0.0   ## noise,  irreproducible
rho02 <- 0.6   ## signal, reproducible
mu01 <-  c(0, 0)
mu02 <- c(3, 3)
var01 <- 1
var02 <- 1
# for x=1
rho11 <- rho01  ## noise,  irreproducible
rho12 <- 0.9  ## signal, reproducible
mu11 <-  mu01
mu12 <-  mu02
var11 <- var01
var12 <- var02

## generate data 
dat <- datSim(n, bp0,
mu01, mu02, var01, var02, rho01, rho02, 
mu11, mu12, var11, var12, rho11, rho12)
}
}
\references{
\itemize{
\item Feipeng Zhang (2017).
A toy example for creating R package.
Qunhua Li's group metting, 12, xxx--xxx.
}
}
\author{
Feipeng Zhang
}
\keyword{myadd}
