% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segCCR.R
\name{segCCR}
\alias{segCCR}
\title{Fit the segmented correspondence curve regression}
\usage{
segCCR(data, par.ini, tm, NB = 100, sig.level = 0.05)
}
\arguments{
\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)
containing the variables in the model. If not found in data,
the variables are taken from environment(formula),
typically the environment from which segCCR is called.}

\item{par.ini}{the initial values for the estimate parameters. The first component is
the change point. If is.null(par.ini) == TRUE, par.ini is set in the
the details.}

\item{tm}{The vector of tm.}

\item{NB}{The bootstrap times to obtain the estimated standard errors.}

\item{sig.level}{The significant level. Default is 0.05.}
}
\value{
A list with the elements:

\describe{
\item{coefficients}{a named vector of coefficients.}
\item{std.error}{the estimated standard errors.}
\item{CI}{the confidence intervals.}
\item{pv}{the p-values for individual test.}
\item{jointpv}{the p-values for joint test.}
}
}
\description{
Fit the correspondence curve regression
}
\details{
Please refer to \strong{Zhang, F. and Li, Q. (2022)}.
}
\examples{
## The example of ChIP-seq data
\dontrun{
data(ChIPseq)
## estimate
m = 100
tm <- seq(0.01, 0.999, length.out = m)
nx = nlevels(factor(ChIPseq$x))
par.ini = c(0.5, 2, 1, rep(0.1, 2*(nx-1)))  # initial value
fit = segCCR(data = ChIPseq,
           par.ini = par.ini,
           tm=tm,
           NB = 5)
}
}
\references{
Zhang, F. and Li, Q. (2022).
Segmented correspondence curve regression for quantifying
covariate effects on the reproducibility of high-throughput experiments.
}
\author{
Feipeng Zhang and Qunhua Li
}
\keyword{segCCR}
