% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segCCR_test.R
\name{segCCR_test}
\alias{segCCR_test}
\title{test the existence of a change point in segCCR}
\usage{
segCCR_test(dat, tm, NB = 100)
}
\arguments{
\item{dat}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame)
containing the variables in the model.}

\item{tm}{The vector of tm.}

\item{NB}{The replicate times to simulate the p-values.}
}
\value{
A list with the elements:

\describe{
\item{pv}{the p-value.}
\item{Tn}{the Tn test statistic.}
\item{Tn.rw}{the Tn test statistic by random weighted method.}
}
}
\description{
Test the existence of a change in segCCR.
}
\details{
Please refer to \strong{Zhang, F. and Li, Q. (2022)}.
}
\examples{
## The example of ChIP-seq data
\dontrun{
data(ChIPseq)
## test
m = 100
tm <- seq(0.01, 0.999, length.out = m)
fit = segCCR_test(ChIPseq,
           tm=tm,
           NB = 10)
}
}
\references{
Zhang, F. and Li, Q. (2022).
Segmented correspondence curve regression for quantifying
covariate effects on the reproducibility of high-throughput experiments.
}
\author{
Feipeng Zhang and Qunhua Li
}
\keyword{segCCR_test}
