% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segccrTest.R
\name{segccrTest}
\alias{segccrTest}
\title{Test the existence of change point in
the segmented correspondence curve regression model.}
\usage{
segccrTest(y, tm = seq(0.01, 0.999, length.out = 50), NB = 5,
  sig.level = 0.05)
}
\arguments{
\item{y}{an matrix with two columns.}

\item{tm}{The vector of tm.}

\item{NB}{The bootstrap times to obtain p-values.}

\item{sig.level}{The significant level. Default is 0.05.}
}
\value{
A list with the elements:

\describe{
  \item{pv}{p-values.}
  }
}
\description{
Test the existence of change point in
 the correspondence curve regression.
}
\details{
Please refer to \strong{Zhang, F. and Li, Q. (2018)}.
}
\examples{
## The example of salmon data
\dontrun{
data(data_salmon)
## estimate
segccrTest(cbind(data_salmon$spawners, data_salmon$recruits), 
           tm=seq(0.001, 0.999, length.out = 28), 
           NB = 2)
 }   
}
\references{
Zhang, F. and Li, Q. (2018). 
Segmented correspondence curve regression model for quantifying 
reproducibility of high-throughput experiments.  Submitted.
}
\author{
Feipeng Zhang and Qunhua Li
}
\keyword{ccrTest}
